Add 104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int i;
-	unsigned long n1 = 1, n2 = 2, n3;
+	unsigned int n1 = 1, n2 = 2, n3;
 
 	printf("%ul, %ul, ", n1, n2);
 	for (i = 3; i <= 98; i++)
